+++ zzimtory/SnsLogin/LoginViewController.swift
@@ -8,6 +8,8 @@
 import UIKit
 import RxSwift
 import RxCocoa
+import CryptoKit
+import AuthenticationServices
 
 final class LoginViewController: UIViewController {
     private var loginView: LoginView?
@@ -45,4 +47,27 @@ final class LoginViewController: UIViewController {
             .disposed(by: disposeBag)
     }
     
+}
+
+extension LoginViewController {
+    private func randomNonceString(length: Int = 32) -> String {
+        precondition(length > 0)
+        var randomBytes = [UInt8](repeating: 0, count: length)
+        let errorCode = SecRandomCopyBytes(kSecRandomDefault, randomBytes.count, &randomBytes)
+        if errorCode != errSecSuccess {
+            fatalError(
+                "Unable to generate nonce. SecRandomCopyBytes failed with OSStatus \(errorCode)"
+            )
+        }
+        
+        let charset: [Character] =
+        Array("0123456789ABCDEFGHIJKLMNOPQRSTUVXYZabcdefghijklmnopqrstuvwxyz-._")
+        
+        let nonce = randomBytes.map { byte in
+            // Pick a random character from the set, wrapping around if needed.
+            charset[Int(byte) % charset.count]
+        }
+        
+        return String(nonce)
+    }
 }
